resource: serialization
sound serialization
