+++ Engine/egNormalMap.hpp
@@ -19,40 +19,13 @@ namespace Engine::Resources
 
 		void Load_INTERNAL() override;
 		void Unload_INTERNAL() override;
+	protected:
+		NormalMap() : Texture({}){}
+
+	private:
+		SERIALIZER_ACCESS
+
 	};
-
-	inline void NormalMap::Initialize()
-	{
-		Texture::Initialize();
-	}
-
-	inline void NormalMap::PreUpdate(const float& dt)
-	{
-		Texture::PreUpdate(dt);
-	}
-
-	inline void NormalMap::Update(const float& dt)
-	{
-		Texture::Update(dt);
-	}
-
-	inline void NormalMap::PreRender(const float dt)
-	{
-		Texture::PreRender(dt);
-	}
-
-	inline void NormalMap::Render(const float dt)
-	{
-		GetRenderPipeline().BindResource(SR_NORMAL_MAP, m_texture_view_.Get());
-	}
-
-	inline void NormalMap::Load_INTERNAL()
-	{
-		Texture::Load_INTERNAL();
-	}
-
-	inline void NormalMap::Unload_INTERNAL()
-	{
-		Texture::Unload_INTERNAL();
-	}
 }
+
+BOOST_CLASS_EXPORT_KEY(Engine::Resources::NormalMap)
